# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Core Topology Generator Web API
  version: "2026-03-05"
  description: >-
    OpenAPI specification for the CORE topology generator web backend. The API
    combines JSON endpoints used by the web UI with a set of HTML form
    submissions and download helpers. Most routes require an authenticated
    session established via `/login` and maintained with the Flask session
    cookie.
servers:
  - url: http://localhost:9090
    description: Default development server
  - url: https://localhost
    description: Reverse-proxied TLS endpoint (see README for nginx helper)
...
security:
  - sessionCookie: []
paths:
...
  /participant-ui:
    get:
      tags:
        - Participant UI
        - UI Pages
      summary: Render the Participant Console page
      description: >-
        Renders the Participant UI HTML page.
        Initial scenario selection priority is: incoming `?scenario=...` request,
        then locally remembered last selection, then first listed scenario.
      parameters:
        - name: scenario
          in: query
          required: false
          schema:
            type: string
          description: Optional scenario display name or normalized label hint.
      responses:
        '200':
          description: Participant UI HTML.
          content:
            text/html:
              schema:
                type: string
  /participant-ui/gateway:
    get:
      tags:
        - Participant UI
      summary: Resolve nearest gateway for a participant scenario
      parameters:
        - name: scenario
          in: query
          required: false
          schema:
            type: string
          description: Scenario label (display or normalized). Defaults to current selected scenario.
      responses:
        '200':
          description: Gateway lookup result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantUiGatewayResponse'
        '403':
          description: Scenario not assigned for restricted users.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /participant-ui/details:
    get:
      tags:
        - Participant UI
      summary: Participant UI scenario status/details (JSON)
      parameters:
        - name: scenario
          in: query
          required: false
          schema:
            type: string
          description: Scenario label (display or normalized). Defaults to the currently selected scenario.
      responses:
        '200':
          description: Scenario details payload used by the Participant UI dashboard.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantUiDetailsResponse'
  /participant-ui/topology:
    get:
      tags:
        - Participant UI
      summary: Participant UI topology graph (JSON)
      description: >-
        Returns graph nodes/links derived from the most recent saved CORE session XML for the requested scenario.
        Network nodes may be returned as `type: switch` for styling purposes; HITL network/interface nodes are
        returned with `type: hitl` and `is_hitl: true`.
      parameters:
        - name: scenario
          in: query
          required: false
          schema:
            type: string
          description: Scenario label (display or normalized). Defaults to the currently selected scenario.
      responses:
        '200':
          description: Topology graph payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantUiTopologyResponse'
        '403':
          description: Scenario not assigned for restricted users.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /participant-ui/stats:
    get:
      tags:
        - Participant UI
      summary: Retrieve Participant UI open-count stats
      parameters:
        - name: scenario
          in: query
          required: false
          schema:
            type: string
          description: Scenario label (display or normalized). Defaults to current selected scenario.
      responses:
        '200':
          description: Scenario and totals open-count telemetry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantUiStatsResponse'
  /participant-ui/record-open:
    post:
      tags:
        - Participant UI
      summary: Record that a participant link was opened
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipantUiRecordOpenRequest'
      responses:
        '200':
          description: Open count recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantUiRecordOpenResponse'
        '400':
          description: Missing or invalid href.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /participant-ui/open:
    get:
      tags:
        - Participant UI
      summary: Redirect to the participant URL for a scenario
      parameters:
        - name: scenario
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to participant URL.
        '404':
          description: No participant URL available for the resolved scenario.
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask session cookie issued after authenticating with `/login`.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          nullable: true
        error:
          type: string
...
    ParticipantUiGraphInterface:
      type: object
      description: Best-effort interface details extracted from CORE session XML.
      properties:
        name:
          type: string
          nullable: true
        mac:
          type: string
          nullable: true
        ipv4:
          type: string
          nullable: true
        ipv4_mask:
          type: string
          nullable: true
        ipv6:
          type: string
          nullable: true
        ipv6_mask:
          type: string
          nullable: true
      additionalProperties: true
    ParticipantUiGraphNode:
      type: object
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
          description: Node identifier (device or network id from CORE XML).
        name:
          type: string
        type:
          type: string
          description: Node category (e.g., router/host/switch/hitl).
        services:
          type: array
          items:
            type: string
        interfaces:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantUiGraphInterface'
        ipv4s:
          type: array
          items:
            type: string
        subnets:
          type: array
          items:
            type: string
        is_vulnerability:
          type: boolean
        is_hitl:
          type: boolean
      additionalProperties: true
    ParticipantUiGraphLink:
      type: object
      required:
        - node1
        - node2
      properties:
        node1:
          type: string
        node2:
          type: string
        node1_name:
          type: string
          nullable: true
        node2_name:
          type: string
          nullable: true
      additionalProperties: true
    ParticipantUiTopologyResponse:
      type: object
      required:
        - ok
        - nodes
        - links
        - subnets
        - vulnerability_ips
      properties:
        ok:
          type: boolean
        scenario_norm:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantUiGraphNode'
        links:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantUiGraphLink'
        subnets:
          type: array
          items:
            type: string
        vulnerability_ips:
          type: array
          items:
            type: string
      additionalProperties: true
    ParticipantUiDetailsResponse:
      type: object
      required:
        - ok
        - scenario_norm
        - scenario
        - gateway
        - open_stats
        - execute
        - session
        - counts
        - subnetworks
        - vulnerability_ips
      properties:
        ok:
          type: boolean
        scenario_norm:
          type: string
        scenario:
          type: object
          required:
            - participant_link_configured
          properties:
            display:
              type: string
              nullable: true
            assigned:
              type: boolean
            placeholder:
              type: boolean
            participant_link_configured:
              type: boolean
          additionalProperties: true
        gateway:
          type: string
          description: Best-effort gateway IPv4 address (without CIDR mask).
        open_stats:
          type: object
          properties:
            open_count:
              type: integer
            last_open_ts:
              type: string
              nullable: true
          additionalProperties: true
        execute:
          type: object
          properties:
            last_execute_ts:
              type: string
              nullable: true
            returncode:
              type: integer
              nullable: true
            ok:
              type: boolean
              nullable: true
          additionalProperties: true
        session:
          type: object
          properties:
            session_id:
              type: integer
              nullable: true
            running:
              type: boolean
              nullable: true
            state:
              type: string
              nullable: true
          additionalProperties: true
        counts:
          type: object
          properties:
            nodes:
              type: integer
              nullable: true
            routers:
              type: integer
              nullable: true
            switches:
              type: integer
              nullable: true
            vulnerabilities:
              type: integer
              nullable: true
          additionalProperties: true
        subnetworks:
          type: array
          items:
            type: string
        vulnerability_ips:
          type: array
          items:
            type: string
      additionalProperties: true
    ParticipantUiGatewayResponse:
      type: object
      required:
        - ok
        - scenario_norm
        - nearest_gateway
      properties:
        ok:
          type: boolean
        scenario_norm:
          type: string
        nearest_gateway:
          type: string
    ParticipantUiStatsResponse:
      type: object
      required:
        - ok
        - scenario_norm
        - scenario
        - totals
      properties:
        ok:
          type: boolean
        scenario_norm:
          type: string
        scenario:
          type: object
          properties:
            open_count:
              type: integer
            last_open_ts:
              type: string
              nullable: true
          additionalProperties: true
        totals:
          type: object
          properties:
            open_count:
              type: integer
            last_open_ts:
              type: string
              nullable: true
          additionalProperties: true
    ParticipantUiRecordOpenRequest:
      type: object
      required:
        - href
      properties:
        scenario_norm:
          type: string
          nullable: true
        href:
          type: string
    ParticipantUiRecordOpenResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
        scenario_norm:
          type: string
          nullable: true
        last_open_ts:
          type: string
          nullable: true
        totals:
          type: object
          additionalProperties: true
        scenario:
          type: object
          additionalProperties: true